+++ packages/deployer/test/integration/custom-proxy.test.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const { loadEnvironment, deployOnEnvironment } = require('./helpers/use-environment');
+const { loadEnvironment, deployOnEnvironment } = require('../helpers/use-environment');
 const { rejects } = require('assert/strict');
 const { ContractValidationError } = require('../../internal/errors');
 const { copyFile } = require('fs/promises');
